Area report results are sorted and extended only once, when data is received from API
