+++ src/Forvaltningsportalen/FeatureInfo/PolygonDetailed.js
@@ -3,7 +3,6 @@ import { ListItem, ListItemIcon, ListItemText, Badge } from "@material-ui/core";
 import { KeyboardBackspace } from "@material-ui/icons";
 import "../../style/infobox.css";
 import CustomIcon from "../../Common/CustomIcon";
-import { getTextAreaReport } from "../../Funksjoner/translateAreaReport";
 import PolygonDetailedDescription from "./PolygonDetailedDescription";
 
 const PolygonDetailed = ({
@@ -13,7 +12,6 @@ const PolygonDetailed = ({
   totalArea
 }) => {
   const [numberResults, setNumberResults] = useState(0);
-  const [sortedResult, setSortedResult] = useState(0);
   const detailResultJSON = JSON.stringify(detailResult);
 
   const areaToPresent = item => {
@@ -34,50 +32,9 @@ const PolygonDetailed = ({
   };
 
   useEffect(() => {
-    let sorted = detailResult.sort((a, b) => {
-      return b.km2 - a.km2;
-    });
-    if (resultLayer.code === "MAT") {
-      sorted = sorted.map(item => {
-        return {
-          ...item,
-          navn: item.kode,
-          kode: ""
-        };
-      });
-    }
-    if (resultLayer.code === "N13") {
-      sorted = sorted.map(item => {
-        // NOTE: A08 has been moved to A11, but data still returns A08
-        return {
-          ...item,
-          navn: getTextAreaReport(
-            "N13",
-            item.kode === "A08" ? "A11" : item.kode,
-            "name"
-          ),
-          beskrivelse: getTextAreaReport(
-            "N13",
-            item.kode === "A08" ? "A11" : item.kode,
-            "description"
-          )
-        };
-      });
-    }
-    if (resultLayer.code === "ANF") {
-      sorted = sorted.map(item => {
-        return {
-          ...item,
-          navn: item.kode,
-          kode: ""
-          // beskrivelse: getTextAreaReport("ANF", item.kode, "description")
-        };
-      });
-    }
-    const number = sorted ? sorted.length : 0;
-    setSortedResult(sorted);
+    const number = detailResult ? detailResult.length : 0;
     setNumberResults(number);
-  }, [detailResult, detailResultJSON, resultLayer]);
+  }, [detailResult, detailResultJSON]);
 
   const iconSize = icon => {
     if (icon && ["terrain", "flag"].includes(icon))
@@ -127,8 +84,8 @@ const PolygonDetailed = ({
           />
         </ListItem>
         <div className="infobox-details-wrapper">
-          {sortedResult &&
-            sortedResult.map((item, index) => {
+          {detailResult &&
+            detailResult.map((item, index) => {
               return (
                 <div
                   key={index}
